+++ Sources/Core/Service/Transport/TransportResult.swift
@@ -7,55 +7,23 @@
 
 import Foundation
 
-/// Represents an error which can be constructed from a `TransportFailure`.
-public protocol TransportFailureRepresentable: Swift.Error {
-
-    init(transportFailure: TransportFailure)
-
-    var failureResponse: HTTP.Response? { get }
-    var transportError: TransportError? { get }
-}
-
 /// Represents an error that occurred when executing a `Request` using a `TransportService`.
 public struct TransportError: Error, Equatable {
     
     // MARK: - Code Subtype
     public enum Code: Equatable {
+        case redirection
         case clientError(HTTP.Status.ClientError)
         case serverError(HTTP.Status.ServerError)
-        case cancelled
-        case noInternetConnection
-        case timedOut
-        case redirection
-        case other(URLError)
-        case unknownError
-
-        
-        // MARK: - Initializer
-        public init(clientError: Error?) {
-            self = (clientError as? URLError).flatMap {
-                switch $0.code {
-                case .cancelled: return .cancelled
-                case .notConnectedToInternet: return .noInternetConnection
-                case .timedOut: return .timedOut
-                default: return .other($0)
-                }
-            } ?? .unknownError
-        }
+        case unknown
     }
     
     // MARK: - Properties
     public let code: Code
-    public let underlyingError: Error?
 
     // MARK: - Initializers
-    public init(code: Code, underlyingError: Error? = nil) {
+    public init(code: Code) {
         self.code = code
-        self.underlyingError = underlyingError
-    }
-    
-    public init(clientError: Error?) {
-        self.init(code: Code(clientError: clientError), underlyingError: clientError)
     }
 
     // MARK: - Equatable
@@ -88,18 +56,10 @@ public struct TransportFailure: Error, Equatable {
     public let response: HTTP.Response?
 
     // MARK: - Initializers
-    public init(error: TransportError, request: HTTP.Request, response: HTTP.Response?) {
+    public init(error: TransportError, request: HTTP.Request, response: HTTP.Response? = nil) {
         self.error = error
         self.request = request
         self.response = response
-    }
-
-    public init(code: TransportError.Code, request: HTTP.Request, response: HTTP.Response?) {
-        self.init(error: TransportError(code: code), request: request, response: response)
-    }
-
-    public init(code: TransportError.Code, response: HTTP.Response) {
-        self.init(code: code, request: response.request, response: response)
     }
 
     public init(error: TransportError, response: HTTP.Response) {
@@ -138,7 +98,7 @@ public extension HTTP.Response {
         case .redirection: return .failure(TransportFailure(error: TransportError(code: .redirection), response: self))
         case .clientError(let clientError): return .failure(TransportFailure(error: TransportError(code: .clientError(clientError)), response: self))
         case .serverError(let serverError): return .failure(TransportFailure(error: TransportError(code: .serverError(serverError)), response: self))
-        case .unknown: return .failure(TransportFailure(error: TransportError(code: .unknownError), response: self))
+        case .unknown: return .failure(TransportFailure(error: TransportError(code: .unknown), response: self))
         }
     }
 }