+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 #### Enhancements
 
-* None
+* convert to async await
 
 ##### Bug Fixes